// src/usecase/customer/DTOs/update.customer.dto.ts
export interface InputUpdateCustomerDto {
  id: string;
  name: string;
  address: {
    street: string;
    city: string;
    number: number;
    zip: string;
  };
}

export interface OutputUpdateCustomerDto {
  id: string;
  name: string;
  active: boolean;
  rewardpoints: number;
  address: {
    street: string;
    city: string;
    number: number;
    zip: string;
  };
}